requirements.txt
pyspark>=3.4.0
kafka-python>=2.0.2